♻️ refactor: CommentSheet 분리 (Jira Ticket : CCC-33)
